// apps/web/src/generated/teamCrests.ts
export type TeamCrestExt = 'svg' | 'png' | 'jpg' | 'jpeg' | 'webp';

export interface TeamCrestFallback {
  src: string;
  page?: string;
}

export const downloadedTeamCrestExts: Record<string, TeamCrestExt> = {
  "ALG": "webp",
  "ARG": "webp",
  "AUS": "webp",
  "AUT": "webp",
  "BEL": "webp",
  "BIH": "webp",
  "BRA": "webp",
  "CAN": "webp",
  "CIV": "webp",
  "COD": "webp",
  "COL": "webp",
  "CPV": "webp",
  "CRO": "webp",
  "CUW": "webp",
  "CZE": "webp",
  "ECU": "webp",
  "EGY": "webp",
  "ENG": "webp",
  "ESP": "webp",
  "FRA": "webp",
  "GER": "webp",
  "GHA": "webp",
  "HAI": "webp",
  "IRN": "webp",
  "IRQ": "webp",
  "JOR": "webp",
  "JPN": "webp",
  "KOR": "webp",
  "KSA": "webp",
  "MAR": "webp",
  "MEX": "webp",
  "NED": "webp",
  "NOR": "webp",
  "NZL": "webp",
  "PAN": "webp",
  "PAR": "webp",
  "POR": "webp",
  "QAT": "webp",
  "RSA": "webp",
  "SCO": "webp",
  "SEN": "webp",
  "SUI": "webp",
  "SWE": "webp",
  "TUN": "webp",
  "TUR": "webp",
  "URU": "webp",
  "USA": "webp",
  "UZB": "webp"
};